models/maker_taker.py
import numpy as np
from sklearn.linear_model import LogisticRegression

def predict_maker_taker(orderbook, order_type):
    """
    Predict maker/taker proportion using a simple logistic regression (stub).
    """
    # For demo: market order is always taker
    if order_type == "market":
        return {"maker": 0.0, "taker": 1.0}
    # Otherwise, assume 50/50
    return {"maker": 0.5, "taker": 0.5}
