requirements.txt
streamlit
websocket-client
numpy
pandas
scikit-learn
